refactor(api): agents mcp 라우트를 전용 모듈로 추출
10개 라우트를 api/agents/mcp.py로 이동. 핸들러 본문·경로·함수명 무변경.
라우트 테이블 characterization 통과(29건 동일, 가림 쌍 0건).

응답 모델 7종(MCPToolSchemaResponse·MCPServerResponse·MCPToolCallRequest/
Response·MCPBatchToolCall{Request,Response}·MCPManagerStatsResponse)도 함께
옮겼다. 서로만 참조하고(MCPServerResponse.tools → MCPToolSchemaResponse 등)
다른 도메인이 쓰지 않아 자기완결적이다. agents.py 는 모델 22개가 파일 앞부분에
모여 있어 도메인별로 분배해야 한다 — projects 처럼 한 구간에 뭉쳐 있지 않다.

ruff 의 F821 이 MCPToolCall·MCPBatchToolCall 누락을 잡았다. 이들은 응답 모델이
아니라 services.mcp_manager 의 내부 도메인 모델이라 심볼 조사에서 빠졌다.
모듈 레벨 import 이전 시 grep 심볼 카운트는 근사치일 뿐이고 린터가 진실원이다.

mcp 라우트는 전부 3세그먼트 이상(/agents/mcp/...)이라 2세그먼트인
/agents/{agent_id} 와 겹치지 않는다 — include 순서가 자유롭다.

게이트: mypy 0, pytest 1348 passed(플레이크 1건 제외).
