+++ src/backend/api/agents/__init__.py
@@ -18,6 +18,7 @@
 테스트 2파일이 ImportError·AttributeError 로 깨진다.
 """
 
+from . import mcp
 from ._legacy import (
     ALLOWED_WORKSPACE_ROOTS,
     _resolve_ocr_runtime,
@@ -26,6 +27,10 @@ from ._legacy import (
     get_allowed_workspace_roots,
     router,
 )
+
+# mcp 라우트는 전부 3세그먼트 이상(`/agents/mcp/...`)이라 2세그먼트인
+# `/agents/{agent_id}` 와 겹치지 않는다 — 등록 순서가 자유롭다.
+router.include_router(mcp.router)
 
 __all__ = [
     "ALLOWED_WORKSPACE_ROOTS",
